requirements.txt
requests~=2.27.1
azure-mgmt-storage
azure-storage-blob
azure-identity
azure-mgmt-resource